fix: removed section tag from card
